+++ src/ml/__init__.py
@@ -1,0 +1,1 @@
+"""EcoLima ML package."""